🐛 Fix: TermsModalPage client 분리
