+++ src/app/(main)/(modal-fallback)/terms/page.tsx
@@ -1,8 +1,14 @@
+import { Suspense } from 'react';
+
 import TermsModalPage from '../../@modal/(.)terms/page';
 
 // 새로고침이나 직접 접근 시에도 TermsModalPage를 재활용
 const TermsStandalonePage = () => {
-  return <TermsModalPage />;
+  return (
+    <Suspense fallback={<div>로딩중...</div>}>
+      <TermsModalPage />
+    </Suspense>
+  );
 };
 
 export default TermsStandalonePage;
