+++ src/app/(main)/@modal/(.)terms/page.tsx
@@ -1,210 +1,12 @@
-import { useSearchParams } from 'next/navigation';
+import { Suspense } from 'react';
 
-// 임시 데이터
-import { Terms } from '@common/components/modals/terms/TermsModal.types';
-import { TermsModalType } from '@common/components/modals/terms/TermsModal.types';
-
-import ModalLayout from '@/common/components/layouts/ModalLayout/ModalLayout.client';
-import TermsModal from '@/common/components/modals/terms/TermsModal.client';
-
-// 임시 데이터f
-const termsData: Record<TermsModalType, Terms> = {
-  [TermsModalType.TERMS_OF_SERVICE]: {
-    title: '이용 약관',
-    content: `## 제 1 장 총칙
-
-### 제 1 조 (목적)
-본 약관은 HaRu(이하 "회사"라 한다)가 제공하는 웹사이트 및 관련 서비스(이하 "서비스")의 이용과 관련하여 회사와 이용자 간의 권리, 의무 및 책임사항 등을 규정하는 것을 목적으로 합니다.
-
-### 제 2 조 (용어의 정의)
-#### 1. "웹사이트"란 서비스를 이용자에게 제공하기 위하여 회사가 운영하는 온라인 플랫폼을 말합니다.
-#### 2. "서비스"란 회사가 제공하는 AI 회의 진행 매니저, SNS 이벤트 어시스턴트, 팀 분위기 트래커 등 각종 자동화 기능을 의미합니다.
-#### 3. "이용자"란 본 약관에 따라 회사가 제공하는 서비스를 이용하는 회원 및 비회원을 말합니다.
-#### 4. "회원"이란 웹사이트에 개인정보를 제공하여 회원등록을 한 자로서, 웹사이트의 정보를 지속적으로 제공받으며, 웹사이트가 제공하는 서비스를 계속적으로 이용할 수 있는 자를 말합니다.
-#### 5. "비회원"이란 회원가입 없이 제한된 범위 내에서 서비스를 이용하는 자를 말합니다.
-
-### 제 3 조 (약관의 공시 및 변경)
-#### 1. 본 약관은 회원가입 화면에 게시하여 공시하며, 회사는 사정변경 및 영업상 중요한 사유가 있을 경우 약관을 변경할 수 있으며 변경된 약관은 공지사항을 통해 공시합니다.
-#### 2. 본 약관 및 차후 회사사정에 따라 변경된 약관은 이용자에게 공시된 시점부터 효력을 갖습니다.
-
-### 제 4 조 (약관 외 준칙)
-본 약관에 명시되지 않은 사항이 전기통신기본법, 전기통신사업법, 정보통신망 이용촉진 및 정보보호에 관한 법률, 전자상거래 등에서의 소비자보호에 관한 법률, 약관의 규제에 관한 법률 등 기타 관계 법령에 규정되어 있을 경우에는 그 규정을 따르도록 합니다.
-
-## 제 2 장 이용계약
-
-### 제 5 조 (이용신청)
-#### 1. 이용신청자가 회원가입 안내에서 본 약관과 개인정보처리방침에 동의하고 '가입하기' 버튼을 누르면 이용신청을 할 수 있습니다.
-
-### 제 6 조 (이용신청의 승낙)
-#### 1. 회사는 원칙적으로 서비스 이용신청을 승낙합니다. 다만, 다음 각 호에 해당하는 경우에는 승낙을 유보하거나 거절할 수 있습니다.
-#### 가. 다른 사람의 명의를 사용하여 신청한 경우  
-#### 나. 이용자 정보를 허위로 기재하여 신청한 경우  
-#### 다. 사회의 안녕질서 또는 미풍양속을 저해할 목적으로 신청한 경우  
-#### 라. 기술적 문제 또는 서비스 용량 부족 등 부득이한 사유가 있는 경우  
-#### 마. 기타 회사가 정한 이용신청 요건이 미비한 경우  
-
-## 제 3 장 계약 당사자의 의무
-
-### 제 7 조 (회사의 의무)
-#### 1. 회사는 웹사이트를 안정적이고 지속적으로 운영할 의무가 있습니다.
-#### 2. 회사는 이용자로부터 제기되는 의견이나 불만이 정당하다고 인정될 경우에는 즉시 처리해야 합니다. 단, 즉시 처리가 곤란한 경우에는 이용자에게 그 사유와 처리일정을 공지사항 또는 전자우편을 통해 통보해야 합니다.
-#### 3. 제1항의 경우 수사상의 목적으로 관계기관의 요청이 있거나 영장 제시가 있는 경우, 기타 관계 법령에 의한 경우는 예외로 합니다.
-
-### 제 8 조 (이용자의 의무)
-#### 1. 이용자는 본 약관 및 회사의 공지사항, 웹사이트 이용안내 등을 숙지하고 준수해야 하며 기타 회사의 업무에 방해되는 행위를 해서는 안 됩니다.
-#### 2. 이용자는 회사의 사전 승인 없이 본 웹사이트를 이용해 어떠한 영리행위도 할 수 없습니다.
-#### 3. 이용자는 본 웹사이트를 통해 얻는 정보를 회사의 사전 승낙 없이 복사, 복제, 변경, 번역, 출판, 방송 및 기타의 방법으로 사용하거나 이를 타인에게 제공할 수 없습니다.
-
-## 제 4 장 서비스 이용
-
-### 제 9 조 (서비스 내용)
-#### 1. 이용자는 본 약관의 규정된 사항을 준수해 웹사이트를 이용합니다.
-#### 2. 본 약관에 명시되지 않은 서비스 이용에 관한 사항은 회사가 정해 공지사항에 게시하거나 또는 별도로 공지하는 내용에 따릅니다.
-
-### 제 10 조 (서비스 이용 제한)
-본 웹사이트 이용 및 행위가 다음 각 항에 해당하는 경우 회사는 해당 이용자의 이용을 제한할 수 있습니다.
-#### 1. 공공질서 및 미풍양속, 기타 사회질서를 해하는 경우
-#### 2. 범죄행위를 목적으로 하거나 기타 범죄행위와 관련된다고 객관적으로 인정되는 경우
-#### 3. 타인의 명예를 손상시키거나 타인의 서비스 이용을 현저히 저해하는 경우
-#### 4. 타인의 의사에 반하는 내용이나 광고성 정보 등을 지속적으로 전송하는 경우
-#### 5. 해킹 및 컴퓨터 바이러스 유포 등으로 서비스의 건전한 운영을 저해하는 경우
-#### 6. 다른 이용자 또는 제3자의 지적재산권을 침해하거나 지적재산권자가 지적 재산권의 침해를 주장할 수 있다고 판단되는 경우
-#### 7. 타인의 아이디 및 비밀번호를 도용한 경우
-#### 8. 기타 관계 법령에 위배되는 경우 및 회사가 이용자로서 부적당하다고 판단한 경우
-
-### 제 11 조 (서비스 중지)
-#### 1. 전기통신사업법 상에 규정된 기간통신 사업자 또는 인터넷 망 사업자가 서비스를 중지했을 경우
-#### 2. 정전으로 서비스 제공이 불가능할 경우
-#### 3. 설비의 이전, 보수 또는 공사로 인해 부득이한 경우
-#### 4. 서비스 설비의 장애 또는 서비스 이용의 폭주 등으로 정상적인 서비스 제공이 어려운 경우
-#### 5. 전시, 사변, 천재지변 또는 이에 준하는 국가비상사태가 발생하거나 발생할 우려가 있는 경우
-
-## 제 5 장 기타
-
-### 제 12 조 (면책 및 손해배상)
-#### 1. 천재지변 또는 이에 준하는 불가항력으로 인하여 서비스를 제공할 수 없는 경우에는 회사의 서비스 제공 책임이 면제됩니다.
-#### 2. 아이디와 비밀번호의 관리 및 이용자의 부주의로 인하여 발생되는 손해 또는 제3자에 의한 부정사용 등에 대한 책임은 이용자에게 있습니다.
-#### 3. 이용자가 본 약관의 규정을 위반함으로써 회사에 손해가 발생하는 경우 이 약관을 위반한 이용자는 회사에 발생한 모든 손해를 배상해야 하며 동 손해로부터 회사를 면책시켜야 합니다.
-
-### 제 13 조 (분쟁의 해결)
-#### 1. 회사와 이용자는 본 사이트 이용과 관련해 발생한 분쟁을 원만하게 해결하기 위하여 필요한 모든 노력을 해야 합니다.
-#### 2. 제1항의 규정에도 불구하고 동 분쟁으로 인하여 소송이 제기될 경우 동 소송은 회사의 본사 소재지를 관할하는 법원의 관할로 합니다.
-
-## 부칙
-### 본 약관은 2025년 7월 1일부터 적용됩니다.`,
-  },
-  [TermsModalType.PRIVACY_POLICY]: {
-    title: '개인정보 처리방침',
-    content: `HaRu(이하 '회사'라 한다)는 개인정보보호법 등 관련 법령에 따라 이용자의 개인정보를 보호하고, 권익을 보호하며, 이와 관련한 고충을 원활하게 처리할 수 있도록 다음과 같은 방침을 수립 및 공개합니다.
-
-#### 1. 개인정보의 처리 목적
-회사는 다음의 목적을 위하여 개인정보를 처리하고 있으며, 다음의 목적 이외의 용도로는 이용하지 않습니다.
-- 이용자 가입 의사 확인 및 본인 식별·인증
-- 회원자격 유지 및 관리
-- 서비스 제공 및 계약 이행
-- 고객상담 및 불만처리
-- 새로운 서비스 개발 및 맞춤형 서비스 제공
-
-#### 2. 개인정보의 처리 및 보유 기간
-회사는 법령에 따른 개인정보 보유·이용기간 또는 정보주체로부터 개인정보를 수집 시에 동의받은 개인정보 보유·이용기간 내에서 개인정보를 처리·보유합니다.
-
-#### 3. 처리하는 개인정보 항목
-회사는 다음의 개인정보 항목을 처리한다.
-
-■ 회원가입 시  
-- 필수항목: 이메일 주소, 비밀번호, 이름, 전화번호  
-- 선택항목: 프로필 사진  
-- 수집목적: 회원 관리, 서비스 제공  
-- 보유기간: 회원 탈퇴 시까지  
-
-■ AI 회의 진행 매니저 서비스 이용 시  
-- 수집항목: 회의 제목, 참석자 정보, 회의록 데이터  
-- 수집목적: 서비스 제공 및 개선  
-- 보유기간: 서비스 목적 달성 시까지  
-
-■ SNS 이벤트 어시스턴트 서비스 이용 시  
-- 수집항목: 이벤트명, 참여자 SNS 계정 정보  
-- 수집목적: 서비스 제공  
-- 보유기간: 이벤트 종료 후 1년  
-
-#### 4. 개인정보의 제3자 제공
-회사는 원칙적으로 이용자의 개인정보를 외부에 제공하지 않습니다. 다만, 다음의 경우에는 예외로 합니다.
-#### - 이용자가 사전에 동의한 경우
-#### - 법령의 규정에 의거하거나, 수사 목적으로 법령에 정해진 절차와 방법에 따라 수사기관의 요구가 있는 경우
-
-#### 5. 개인정보 처리의 위탁
-회사는 서비스 제공을 위해 다음과 같이 개인정보 처리업무를 위탁하고 있습니다.
-- 위탁받는 업체: AWS(Amazon Web Services)
-- 위탁업무 내용: 클라우드 서비스 및 데이터 저장
-- 위탁기간: 서비스 제공 기간
-
-#### 6. 정보주체의 권리·의무 및 행사방법
-정보주체는 회사에 대해 언제든지 다음 각 호의 개인정보 보호 관련 권리를 행사할 수 있습니다.
-- 개인정보 열람 요구
-- 오류 등이 있을 경우 정정·삭제 요구
-- 처리정지 요구
-
-#### 7. 개인정보의 파기
-회사는 개인정보 보유기간의 경과, 처리목적 달성 등 개인정보가 불필요하게 되었을 때에는 지체없이 해당 개인정보를 파기합니다.
-- 파기절차: 이용자가 입력한 정보는 목적 달성 후 별도의 DB에 옮겨져 내부 방침 및 기타 관련 법령에 따라 일정기간 저장된 후 혹은 즉시 파기됩니다.
-- 파기방법: 전자적 파일 형태의 정보는 기록을 재생할 수 없는 기술적 방법을 사용합니다.
-
-#### 8. 개인정보 보호책임자
-회사는 개인정보 처리에 관한 업무를 총괄해서 책임지고, 개인정보 처리와 관련한 정보주체의 불만처리 및 피해구제 등을 위하여 아래와 같이 개인정보 보호책임자를 지정하고 있습니다.
-
-▶ 개인정보 보호책임자  
-- 성명: 황지원  
-- 직책: 대표이사  
-- 연락처: privacy@haru-app.com  
-※ 개인정보 보호 담당부서로 연결됩니다.
-
-#### 9. 개인정보 처리방침 변경
-이 개인정보처리방침은 시행일로부터 적용되며, 법령 및 방침에 따른 변경내용의 추가, 삭제 및 정정이 있는 경우에는 변경사항의 시행 7일 전부터 공지사항을 통하여 고지할 것입니다.
-
-### 본 방침은 2025년 7월 1일부터 시행됩니다.
-`,
-  },
-  [TermsModalType.MARKETING_CONSENT]: {
-    title: '마케팅정보수신',
-    content: `HaRu(이하 '회사'라 한다)는 정보통신망 이용촉진 및 정보보호에 관한 법률에 따라 광고성 정보를 전송하기 위해 수신자의 사전 수신 동의를 받고 있습니다.
-
-#### 1. 수집·이용 목적
-• 회사 및 제휴사의 상품·서비스 안내 및 이용 권유
-• 각종 이벤트 및 프로모션 안내
-• 신규 서비스 및 신상품이나 이벤트 정보 등의 안내
-• 인구통계학적 특성에 따른 서비스 제공 및 광고 게재
-
-#### 2. 수집·이용 항목
-• 이메일 주소
-• 휴대폰 번호
-• 서비스 이용기록, 접속 로그, 쿠키, 접속 IP 정보
-
-#### 3. 보유·이용 기간
-회원 탈퇴 시 또는 마케팅 정보 수신 동의 철회 시까지
-
-#### 4. 동의 거부권 및 불이익
-본 동의는 선택사항이며, 동의를 거부하셔도 회사의 서비스 이용에는 제한이 없습니다. 다만, 마케팅 정보 수신 동의를 거부하시는 경우 관련 마케팅 정보를 받아보실 수 없습니다.
-
-#### 5. 수신거부 방법
-• 수신거부를 원하시는 경우 이메일 하단의 '수신거부' 링크를 클릭하시거나
-• 고객센터(support@haru-app.com)로 연락 주시면 즉시 처리해드립니다.
-
-### 본 동의서는 2025년 7월 1일부터 적용됩니다.`,
-  },
-};
+import TermsModalClient from './TermsModalClient';
 
 const TermsModalPage = () => {
-  const searchParams = useSearchParams();
-  const type = searchParams.get('type');
-
-  const terms = termsData[type as TermsModalType];
-
-  if (!terms) return null;
-
   return (
-    <ModalLayout>
-      <TermsModal onClose={() => history.back()} terms={terms} />
-    </ModalLayout>
+    <Suspense fallback={<div>로딩중...</div>}>
+      <TermsModalClient />
+    </Suspense>
   );
 };
